[READLINE] Cleanup main readline structure

+++ src/readline/cursor_manipulation/move_cursor.c
@@ -15,7 +15,7 @@ static void	move_cursor_left(void)
 		rp()->cur_pos[1]--;
 		rp()->cur_pos[0] = search_last_cur_pos_in_line() + 1;
 	}
-	else if (rp()->cur_pos[0] == rp()->prompt_len && rp()->cur_pos[1] == 1)
+	else if ((size_t)rp()->cur_pos[0] == (size_t)rp()->prompt_len && rp()->cur_pos[1] == 1)
 		return ;
 	else
 		rp()->cur_pos[0]--;
